Put the recent changes in the appropiate files and folders
