+++ Pixel Soup/src/frontend/gameconstants.py
@@ -1,3 +1,16 @@
-SCREEN_WIDTH = 800
-SCREEN_HEIGHT = 600
-SCREEN_TITLE = "Game"
+from pathlib import Path
+
+GAME_PATH = str(Path(__file__).parent.resolve())
+
+SCREEN_WIDTH = 1100
+SCREEN_HEIGHT = 700
+SCREEN_TITLE = "Trio Survivals"
+
+SPRITE_SCALING_PLAYER = 0.2
+
+BOTTOM_VIEWPORT_MARGIN = 50
+TOP_VIEWPORT_MARGIN = 100
+
+PLAYER_MOVEMENT_SPEED = 5
+GRAVITY = 1
+PLAYER_JUMP_SPEED = 20
